新增支持Glide并替代URL
